+++ mystrings/ctype-bin.c
@@ -273,7 +273,7 @@ static int my_wc_mb_bin(const CHARSET_INFO * const cs __attribute__((unused)),
 
 static void my_hash_sort_8bit_bin(const CHARSET_INFO * const cs __attribute__((unused)),
                                   const uchar *key, size_t len,
-                                  ulong *nr1, ulong *nr2)
+                                  uint32_t *nr1, uint32_t *nr2)
 {
   const uchar *pos = key;
   
@@ -294,7 +294,7 @@ static void my_hash_sort_8bit_bin(const CHARSET_INFO * const cs __attribute__((u
 
 static void my_hash_sort_bin(const CHARSET_INFO * const cs __attribute__((unused)),
                              const uchar *key, size_t len,
-                             ulong *nr1, ulong *nr2)
+                             uint32_t *nr1, uint32_t *nr2)
 {
   const uchar *pos = key;
